remove filter lists
Brave's services have already migrated away from lists in this
repository. Anyone looking for a replacement for these lists should use
https://github.com/brave/adblock-resources.
