+++ src/lib.rs
@@ -28,7 +28,6 @@ pub mod blocker;
 pub mod optimizer;
 pub mod url_parser;
 pub mod engine;
-pub mod filter_lists;
 pub mod resources;
 pub mod cosmetic_filter_cache;
 pub mod data_format;
